pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nautilus-monster"
version = "0.1.0"
description = "NautilusMonster V3: AI-driven crypto trading runtime using Coinbase data and Coinbase/Alpaca execution adapters."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "NautilusMonster Team" }]
dependencies = [
  "fastapi>=0.115.12",
  "uvicorn>=0.34.1",
  "pydantic>=2.11.3",
  "PyYAML>=6.0.2",
  "websockets>=13.1,<14.0",
  "aiohttp>=3.11.18",
  "polars>=1.27.1",
  "numpy>=2.2.4",
  "scikit-learn>=1.6.1",
  "hmmlearn>=0.3.3",
  "qdrant-client>=1.14.2",
  "redis>=5.2.1",
  "httpx>=0.28.1",
  "prometheus-client>=0.21.1",
  "streamlit>=1.44.1",
  "mlflow>=2.21.3",
  "prefect>=3.3.4",
  "python-dotenv>=1.1.0",
  "alpaca-py>=0.40.1",
  "coinbase-advanced-py>=1.8.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "ruff>=0.11.5",
  "mypy>=1.15.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = [
  "app*",
  "data_plane*",
  "models*",
  "decision_engine*",
  "risk_engine*",
  "execution*",
  "backtesting*",
  "control_plane*",
  "orchestration*",
  "observability*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B"]
ignore = ["UP042"]
